+++ python_base_app/base_ci_toolbox.py
@@ -111,8 +111,14 @@ PREDEFINED_ENV_VARIABLES = [
     VarStatus('MAX_CPUS', "Maximum number of parallel CPUs/threads to be used during build", "max_cpus"),
 ]
 
-DEFAULT_PYPI_REPOSITORY = "https://pypi.org"
+DEFAULT_PYPI_REPOSITORY = "https://upload.pypi.org/legacy/"
+DEFAULT_TEST_PYPI_REPOSITORY = "https://test.pypi.org/legacy/"
+
+DEFAULT_PYPI_URL_ENV_NAME = "PYPI_API_URL"
+DEFAULT_TEST_PYPI_URL_ENV_NAME = "TEST_PYPI_API_URL"
+
 DEFAULT_PYPI_TOKEN_ENV_NAME = "PYPI_API_TOKEN"
+DEFAULT_PYPI_USER_ENV_NAME = "PYPI_API_USER"
 
 predefined_env_variables = None
 
@@ -131,6 +137,8 @@ default_setup = {
     "ci_stage_teardown": STAGE_TEARDOWN,
     "ci_stage_publish_package": STAGE_PUBLISH_PACKAGE,
     "ci_stage_publish_pypi_package": STAGE_PUBLISH_PYPI_PACKAGE,
+    "ci_stage_build_pip_dependencies": [ "babel" ],
+    "ci_stage_publish_pip_package_pip_dependencies": [ "twine" ],
     "require_teardown": False,
     "bin_dir": "bin",
     "test_dir": "test",
@@ -299,8 +307,10 @@ def get_python_packages(p_main_setup_module, p_arguments, p_include_contrib_pack
             * 1: filename of the python pip package (excluding path)
             * 2: module name
             * 3: vars of the package
-            * 4: target PyPi repository URL
-            * 5: variable name containing token for target PyPi repository
+            * 4: variable name containing the target PyPi repository URL
+            * 5: variable name containing the token for the target PyPi repository
+            * 6: variable name containing the username for the target PyPi repository
+            * 7: default target PyPi repository URL
     """
 
     var = get_vars(p_setup_params=p_main_setup_module.extended_setup_params)
@@ -316,28 +326,42 @@ def get_python_packages(p_main_setup_module, p_arguments, p_include_contrib_pack
     branch = var["setup"].get("GIT_BRANCH")
     branch_target_rep_map = var["setup"]["publish_pypi_package"]
 
-    target_rep = None
+    target_rep_url_env_name = None
     target_rep_token_env_name = None
+    target_rep_user_env_name = None
 
     if isinstance(branch_target_rep_map, dict) and branch is not None:
         target_rep_map_entry = branch_target_rep_map.get(branch)
 
         if target_rep_map_entry is not None:
-            target_rep = target_rep_map_entry[0]
+            target_rep_url_env_name = target_rep_map_entry[0]
             target_rep_token_env_name = target_rep_map_entry[1]
 
-    if target_rep is None:
-        target_rep = DEFAULT_PYPI_REPOSITORY
+            if len(target_rep_map_entry) > 2:
+                target_rep_user_env_name = target_rep_map_entry[2]
+
+    if target_rep_url_env_name is None:
+        target_rep_url_env_name = DEFAULT_PYPI_URL_ENV_NAME
 
     if target_rep_token_env_name is None:
         target_rep_token_env_name = DEFAULT_PYPI_TOKEN_ENV_NAME
+
+    if target_rep_user_env_name is None:
+        target_rep_user_env_name = DEFAULT_PYPI_USER_ENV_NAME
+
+    if target_rep_url_env_name == DEFAULT_TEST_PYPI_URL_ENV_NAME:
+        target_rep_default_url = DEFAULT_TEST_PYPI_REPOSITORY
+
+    else:
+        target_rep_default_url = DEFAULT_PYPI_REPOSITORY
 
     contributing_setup_modules = load_contributing_setup_modules(p_main_setup_module)
 
     python_packages = []
 
     python_packages.append((app_dir, get_python_package_name(p_var=var), var["setup"]["module_name"], var,
-                            target_rep, target_rep_token_env_name))
+                            target_rep_url_env_name, target_rep_token_env_name, target_rep_user_env_name,
+                            target_rep_default_url))
 
     if p_include_contrib_packages:
         for contributing_setup_module in contributing_setup_modules:
@@ -351,7 +375,8 @@ def get_python_packages(p_main_setup_module, p_arguments, p_include_contrib_pack
                 include_path = contrib_dir
 
             python_packages.append((include_path, get_python_package_name(p_var=contrib_var), module_name, contrib_var,
-                                    target_rep, target_rep_token_env_name))
+                                    target_rep_url_env_name, target_rep_token_env_name, target_rep_user_env_name,
+                                    target_rep_default_url))
 
     return python_packages